feat: add standalone theme controls
